backend/requirements.txt
fastapi
uvicorn[standard]
supabase
httpx
python-dotenv
docling
pandas
tabulate
